1，完成“个人”界面个人信息接口调用； 2，增加设置功能（还差清除聊天记录，版本更新）

+++ jixin/app/src/main/java/com/atman/jixin/model/response/LoginResultModel.java
@@ -66,7 +66,7 @@ public class LoginResultModel {
         private long id;
         private String memberName;
         private String memberTrueName;
-        private int atmanUserId;
+        private long atmanUserId;
         private String memberMobile;
         private String memberAvatar;
         private int memberSex;
@@ -118,11 +118,11 @@ public class LoginResultModel {
             this.memberTrueName = memberTrueName;
         }
 
-        public int getAtmanUserId() {
+        public long getAtmanUserId() {
             return atmanUserId;
         }
 
-        public void setAtmanUserId(int atmanUserId) {
+        public void setAtmanUserId(long atmanUserId) {
             this.atmanUserId = atmanUserId;
         }
